指针

+++ src/day02/ConstDemo.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 
-const int PI = 3.1415926;
+const double PI = 3.1415926;
 
 int main() {
